% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cospa}
\alias{cospa}
\title{estimation function for the co-sparse matrix}
\usage{
cospa(X, Y, smin, smax, nrank, alpha = 1, verbose = FALSE)
}
\arguments{
\item{X}{the predictor matrix}

\item{Y}{the response matrix}

\item{smin}{the minimum sparsity level}

\item{smax}{the maximum sparsity level}

\item{nrank}{the rank of the coefficient matrix}

\item{alpha}{the paramter to tune the penalty level, default is 1}

\item{verbose}{indicator to print iteration procedure}
}
\value{
A list includes the estimation result
\itemize{
  \item d - the singular values 
  \item U - the left singular vectors 
  \item V - the right singular vectors 
  \item time - the computation time 
  \item nrank - the estimated rank 
}
}
\description{
estimation function for the co-sparse matrix
}
