% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{init.data}
\alias{init.data}
\title{initialize the data for simulation}
\usage{
init.data(
  n = 100,
  p = 200,
  q = 100,
  d = c(10, 15, 20),
  su = 3,
  sv = 4,
  snr = 0.5,
  corrX = c("AR", "CS"),
  rhoX = 0.5,
  corrE = c("AR", "I", "CS"),
  rhoE = 0.3
)
}
\arguments{
\item{n}{sample size of the observation}

\item{p}{dimension of predictors}

\item{q}{dimension of responses}

\item{d}{singular values of the coefficient matrix}

\item{su}{sparsity level of left singular vectors}

\item{sv}{sparsity level of right singular vectors}

\item{snr}{signal-to-noise rate for the coefficient matrix}

\item{corrX}{covariance matrix type of the predictors}

\item{rhoX}{the parameter for the predictor covariance matrix}

\item{corrE}{covariance matrix type of the noise vector}

\item{rhoE}{the parameter for the noise covariance matrix}
}
\description{
initialize the data for simulation
}
